introducing BaseProxyHandler abstraction based on review feedback
