+++ src/shared.ts
@@ -11,6 +11,7 @@ const {
     getOwnPropertySymbols,
     preventExtensions,
     hasOwnProperty,
+    freeze,
 } = Object;
 
 const {
@@ -34,6 +35,7 @@ export {
     getOwnPropertySymbols,
     preventExtensions,
     hasOwnProperty,
+    freeze,
 };
 
 const OtS = {}.toString;
